More scene test separation
